modules/web-app/src/test/java/com/example/gomesrodris/archburgers/adapters/presenters/ItemCardapioPresenterTest.java
package com.example.gomesrodris.archburgers.adapters.presenters;//import static org.junit.jupiter.api.Assertions.*;

import com.example.gomesrodris.archburgers.adapters.dto.ItemCardapioDto;
import com.example.gomesrodris.archburgers.adapters.dto.ValorMonetarioDto;
import com.example.gomesrodris.archburgers.adapters.testUtils.TestLocale;
import com.example.gomesrodris.archburgers.domain.entities.ItemCardapio;
import com.example.gomesrodris.archburgers.domain.valueobjects.TipoItemCardapio;
import com.example.gomesrodris.archburgers.domain.valueobjects.ValorMonetario;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class ItemCardapioPresenterTest {
    @BeforeAll
    static void beforeAll() {
        TestLocale.setDefault();
    }

    @Test
    void entityToPresentationDto() {
        var dto = ItemCardapioPresenter.entityToPresentationDto(new ItemCardapio(123,
                TipoItemCardapio.LANCHE, "Cheeseburger", "Hambúrguer com queijo", new ValorMonetario("25.99")));

        assertThat(dto).isEqualTo(new ItemCardapioDto(
                123, "LANCHE", "Cheeseburger", "Hambúrguer com queijo",
                new ValorMonetarioDto("25.99", "R$ 25,99")
        ));
    }

}